Separate Permission Logic Handling
